+++ src/compiler/parser.c
@@ -63,13 +63,13 @@ void risa_parser_error_at(RisaParser* parser, RisaToken token, const char* msg) 
         size_t ln;
         size_t col;
 
-        risa_lib_mem_lncol(parser->lexer.source, token.index, &ln, &col);
+        risa_lib_mem_lncol(parser->lexer.source, (size_t) token.index, &ln, &col);
         RISA_ERROR(parser->io, "at %zu:%zu in script: %s\n", ln, col, msg);
     } else {
         size_t ln;
         size_t col;
 
-        risa_lib_mem_lncol(parser->lexer.source, token.index, &ln, &col);
+        risa_lib_mem_lncol(parser->lexer.source, (size_t) token.index, &ln, &col);
         RISA_ERROR(parser->io, "at %zu:%zu in script: Invalid token\n", ln, col);
     }
